+++ functions/actions/candidates/checkSignInMethodsForEmail.js
@@ -1,13 +1,13 @@
 export default (auth) => {
 
-  return fetchSignInMethodsForEmail;
+  return checkSignInMethodsForEmail;
 
   /**
    * Check if a user of given email address is enabled
    * @param {string} email 
    * @returns 
    */
-  async function fetchSignInMethodsForEmail({ email }) {
+  async function checkSignInMethodsForEmail({ email }) {
     try {
       const user = await auth.getUserByEmail(email);
       return !user.disabled && user.providerData.length && !!user.providerData.find(provider => provider.providerId === 'password');
